Simplify list layout
